refactor(roadmap): files refactored
The Roadmap component was updated to use SCSS nested properties
use REM units instead of pixels
Additionally, the component file was moved to a Roadmap folder within the components folder, and import lines were updated to use local paths instead of global ones. Finally, some functions were refactored to use arrow function syntax
