+++ src/assets/index.scss
@@ -132,19 +132,26 @@ td {
   }
 }
 
+.roadmap-boxes {
+  .field-title {
+    font-family: 'Comfortaa', Arial, sans-serif;
+  }
 
-.roadmap-container{
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
+  .roadmap-container {
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+  }
 }
 
-.card-icon{
+.card-icon {
   height: 70%;
   width: 70%;
 }
 
-.field-title,.roadmap-card, .subItem-title {
-  font-family: 'Comfortaa', Arial, sans-serif;
+.roadmap-card {
+  .roadmap-card-item {
+    font-family: 'Comfortaa', Arial, sans-serif;
+  }
 }
 
 .bg-247 {
@@ -158,31 +165,19 @@ td {
   background-color: rgb(230, 225, 236);
 }
 
-.progress-bar {
-  font-family: 'Comfortaa', Arial, sans-serif;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100px;
-  height: 100px;
-  border-radius: 50%;
-  color: #5E5E5E;
-  background: 
-    radial-gradient(closest-side, rgb(224 242 254) 79%, transparent 80% 100%),
-    conic-gradient(#6DA5D8 85%, #D1D5DB 0);    
-  
-}
+.concepts {
+  .subItem-title {
+    color: #5e5e5e;
+    font-family: 'Comfortaa', Arial, sans-serif;
+  }
+  .subItem-subTitle {
+    color: #5e5e5e;
+  }
 
-
-.subItem-title, .subItem-subTitle{
-  color: #5E5E5E;
-}
-
-.bottom-navigation{
-  position: relative;
-  bottom: 25px;
-  left: 30px;
- 
-
- 
+  .progress-bar {
+    font-family: 'Comfortaa', Arial, sans-serif;
+    color: #5e5e5e;
+    background: radial-gradient(closest-side, rgb(224 242 254) 79%, transparent 80% 100%),
+      conic-gradient(#6da5d8 85%, #d1d5db 0);
+  }
 }
